start refactoring using path aliases
