+++ package.json
@@ -49,7 +49,6 @@
     "@typescript-eslint/eslint-plugin": "~5.46",
     "@typescript-eslint/parser": "~5.46",
     "babel-jest": "~29.6",
-    "babel-loader": "~9.1",
     "clean-webpack-plugin": "~4.0",
     "core-js": "~3.32",
     "css-loader": "~6.8",
@@ -81,6 +80,7 @@
     "style-loader": "~3.3",
     "svg-inline-loader": "~0.8",
     "ts-jest": "~29.1",
+    "ts-loader": "^9.5.0",
     "typedoc": "~0.24",
     "typescript": "~5.1",
     "url-loader": "~4.1",
