+++ src/components/Explore/LeftSidebar/SpeciesDetailView.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useContext, useEffect, useState } from 'react';
+import React, { FC, useContext } from 'react';
 import Typography from '@mui/material/Typography';
 import { Box, Button, Chip, Stack, capitalize } from '@mui/material';
 import { ArrowBackOutlined, FileDownloadOutlined, ScienceOutlined } from '@mui/icons-material';
