+++ src/components/Explore/LeftSidebar/Station/Text.tsx
@@ -1,13 +1,13 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Alert from '@mui/material/Alert';
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import { Button } from '@mui/material';
 import { OpenInNewOutlined } from '@mui/icons-material';
-import hathiTrustLogoSvg from '../../images/hathi_trust.svg';
-import Loading from '../Loading';
-import { theme } from '../../theme';
+import hathiTrustLogoSvg from '@app/images/hathi_trust.svg';
+import Loading from '@app/components/Loading';
+import { theme } from '@app/theme';
 
 interface Props {
     station: StationDetails | null;
