+++ src/components/Explore/LeftSidebar/StationDetailView.tsx
@@ -6,10 +6,10 @@ import Typography from '@mui/material/Typography';
 import { DataStateContext } from '../../../store/contexts';
 import { useStationDetails } from '../../../utils/hooks';
 import TabsGroup from '../../TabsGroup';
-import StationDetails from '../../Station/Details';
-import StationEnvironment from '../../Station/Environment';
-import StationSpecies from '../../Station/Species';
-import StationText from '../../Station/Text';
+import StationDetails from './Station/Details';
+import StationEnvironment from './Station/Environment';
+import StationSpecies from './Station/Species';
+import StationText from './Station/Text';
 import { theme } from '../../../theme';
 import { ArrowDropDownOutlined, FileDownloadOutlined } from '@mui/icons-material';
