+++ src/types/global.d.ts
@@ -52,10 +52,6 @@ declare module '*.webp' {
 }
 
 declare module '*.svg' {
-    import * as React from 'react';
-
-    export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-
     const src: string;
     export default src;
 }
