+++ tsconfig.json
@@ -1,11 +1,15 @@
 {
   "compilerOptions": {
+    "outDir": "dist",
     "target": "es6",
     "lib": [
       "dom",
       "dom.iterable",
       "esnext"
     ],
+    "paths": {
+      "@app/*": ["./src/*"],
+    },
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -16,7 +20,6 @@
     "moduleResolution": "node",
     "isolatedModules": true,
     "resolveJsonModule": false,
-    "noEmit": true,
     "jsx": "react",
     "sourceMap": true,
     "declaration": true,
